requirements.txt
numpy==1.26.4
Django==4.2.6
djangorestframework==3.14.0
python-dotenv==0.20.0
psycopg2-binary==2.9.9
